Split file size analysis from Lighthouse verifications
