+++ web/scripts/verify-build-size.js
@@ -1,0 +1,53 @@
+#!/usr/bin/env node
+
+const fs = require('fs')
+const path = require('path')
+const walk = require('walk')
+const chalk = require('chalk')
+
+// A number denoting maximum file size in bytes.
+const FILE_SIZE_LIMIT = parseInt(process.env.file_size_limit || process.env.npm_package_config_file_size_limit)
+
+let failure = false
+
+/**
+* Traverse the build folder and verify that built files are smaller than a 
+* defined threshold.
+*/
+let options = {
+    listeners: {
+        file: function (root, fileStats, next) {
+            let filePath = path.join(root, fileStats.name)
+            let fileStat = fs.statSync(filePath)
+            if (fileStat.size > FILE_SIZE_LIMIT) {
+                failure = true
+                console.log(chalk.red(`${filePath} is ${fileStat.size} bytes. It is too big!\n`))
+            }
+            next()
+        },
+        errors: function (root, nodeStatsArray, next) {
+            next()
+        },
+        end: function() {
+            if (failure === false) {
+                console.log(`Success! All build files are below the size threshold.`)
+            } else {
+                console.log(chalk.red(`Run 'npm run analyze-build' to see what is contributing to large files.`))
+            }
+        }
+    }
+}
+
+if (fs.existsSync('build')) {
+    console.log(`Verifying individual file sizes in the build are less than ${FILE_SIZE_LIMIT} bytes...`)
+    console.log(process.env.npm_package_config_file_size_limit)
+    walk.walkSync('build', options);
+} else {
+    console.log(`Run 'npm prod:build' to generate a build.`)
+}
+
+if (failure) {
+    process.exit(1)
+} else {
+    process.exit(0)
+}
